removed rectangles from Canvas
