+++ client/components/Rectangle.js
@@ -1,0 +1,42 @@
+import React, {Component} from 'react'
+
+class Rectangle extends Component {
+  constructor() {
+    super()
+    this.state = {
+      rectangles: [
+        {
+          x: 10,
+          y: 10,
+          width: 100,
+          height: 100,
+          fill: 'red',
+          name: 'rect1'
+        },
+        {
+          x: 150,
+          y: 150,
+          width: 100,
+          height: 100,
+          fill: 'green',
+          name: 'rect2'
+        }
+      ]
+    }
+  }
+  render() {
+    return (
+      <Rect
+        x={this.props.x}
+        y={this.props.y}
+        width={this.props.width}
+        height={this.props.height}
+        fill={this.props.fill}
+        name={this.props.name}
+        draggable
+      />
+    )
+  }
+}
+
+export default Rectangle
